added a grid view to buffet , fixed spacing for smaller window sizes made magnify and grid view to individual components and also added a layout selector component
